chore[st]: moving challenge dependencies into challenge model
