+++ SmartTutor/ContentModel/LearningObjects/ChallengeModel/MetricRules/MetricRangeRule.cs
@@ -1,7 +1,7 @@
 ﻿using RepositoryCompiler.CodeModel.CaDETModel.CodeItems;
 using System.Collections.Generic;
 
-namespace SmartTutor.ContentModel.LearningObjects.MetricRules
+namespace SmartTutor.ContentModel.LearningObjects.ChallengeModel.MetricRules
 {
     public class MetricRangeRule
     {
